refactor: move test files
